Refactored the code to reduce code duplication This was achieved through the use of the maybe_async crate
