+++ src/lib.rs
@@ -5,10 +5,17 @@ LICENSE: BSD3 (see LICENSE file)
 
 #![no_std]
 
-pub mod interface;
-pub mod wrapper;
-#[cfg(feature = "async")]
-pub mod wrapper_async;
+#[cfg(all(feature = "sync", feature = "async"))]
+compile_error!("features `sync` and `async` are mutually exclusive");
+
+#[cfg(not(any(feature = "sync", feature = "async")))]
+compile_error!("either `sync` or `async` must be enabled");
+
+#[cfg(any(
+    all(feature = "sync", not(feature = "async")),
+    all(feature = "async", not(feature = "sync"))
+))]
+pub mod interface; pub mod wrapper;
 
 /// Errors in this crate
 #[derive(Debug)]
